clean:テスト対象のクラスの変更を受けてテストも変更
